Adding transactions to the simulation library

+++ ln_ms_lib/src/sim_node.rs
@@ -5,6 +5,6 @@ use serde::{Serialize, Deserialize};
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct SimNode {
     pub name: String,
-    pub initial_balance: i32,
+    pub initial_balance: u64,
     pub running: bool
 }
